Create a cmponent for homepage music card
