+++ src/components/musicNft/musicNft.module.css
@@ -1,0 +1,49 @@
+.controls {
+  display: flex;
+  justify-content: center;
+  gap: 5px;
+  margin-top: 50px;
+}
+
+.prevButton, .nextButton {
+  border: none;
+  background: #000;
+  padding: 10px;
+  color: white;
+  cursor: pointer;
+  font-size: 16px;
+}
+
+.prevButton:hover, .nextButton:hover {
+  background: #333;
+}
+
+.gridContainer{
+  overflow: hidden;
+  width: 100%;
+  cursor: pointer;
+  padding-top: 1rem;
+}
+.nftGrid {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 25px;
+  max-width: 1200px;
+  width: 100%;
+
+}
+
+.nftGrid p {
+  margin-top: 10px;
+}
+
+.nftGrid a {
+  color: white;
+  cursor: pointer;
+}
+
+.nftGrid a:hover {
+  color: var(--prim-color);
+}
