+++ src/pages/homepage/homepage.module.css
@@ -2,28 +2,43 @@
   display: flex;
   justify-content: center;
   padding: 20px;
-  margin-left: 250px; 
 }
 
 .mainContent {
   width: 100%;
-  background-image: url("/assets/dreokt.jpg");
-  background-size: 100% 520px;
-  background-repeat: no-repeat;
   min-height: 100vh;
   font-family: var(--prim-ROBOTO);
   color: var(--prim-color-2);
 }
 
 .header {
+ position: relative;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
   text-align: center;
   margin-bottom: 3rem;
-  height: 200px;
-   background: linear-gradient(270deg, #601baa93, rgba(0, 0, 0, .5), #4a147398);
-  background-size: 300% 100%;
-   animation: gradientAnimation 8s ease infinite alternate;
+  height: 300px;
+  transition: background-image 1.5s ease-in-out;
+  overflow: hidden;
+}
+
+.headerOverlay {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 100%;
   padding: 2.85rem 2rem;
-  
+  background: linear-gradient(270deg, #601baa93, rgba(0, 0, 0, 0.7), #4a147398);
+  background-size: 200% 200%;
+  animation: gradientAnimation 8s ease infinite alternate;
+}
+
+.fading {
+  opacity: 0.8;
+  transition: opacity 1s ease;
 }
 
 
@@ -32,64 +47,41 @@
   font-weight: bold;
   font-family: var(--prim-ORBITRON);
   margin-bottom: .5rem;
+   animation: fadeInUp 1s ease-out forwards;
+    text-shadow: 0 0 15px rgba(255, 255, 255, 0.6);
 }
 
 .header p {
   font-size: 1.2rem;
   opacity: 0.8;
+   text-shadow: 0 0 10px rgba(255, 255, 255, 0.5);
 }
 
-.controls {
-  display: flex;
-  justify-content: center;
-  gap: 5px;
-  margin-top: 50px;
-}
-
-.prevButton, .nextButton {
-  border: none;
-  background: #000;
-  padding: 10px;
-  color: white;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.prevButton:hover, .nextButton:hover {
-  background: #333;
-}
-
-.gridContainer{
-  overflow: hidden;
+.header::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
   width: 100%;
-  cursor: pointer;
-  padding-top: 1rem;
-}
-.nftGrid {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 25px;
-  max-width: 1200px;
-  width: 100%;
-
+  height: 100%;
+  background-image: url("/assets/particle-overlay.png");
+  background-size: cover;
+  opacity: 0.4;
+  animation: floatParticles 20s linear infinite;
+  pointer-events: none;
 }
 
-.nftGrid p {
-  margin-top: 10px;
+@keyframes gradientAnimation {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
 }
-
-.nftGrid a {
-  color: white;
-  cursor: pointer;
-}
-
-.nftGrid a:hover {
-  color: var(--prim-color);
-}
-
-
 
 @keyframes fadeInUp {
   from {
@@ -102,7 +94,22 @@
   }
 }
 
-/* @keyframes gradientAnimation {
-  0% { background-position: 0% 0; }
-  100% { background-position: 100% 0; }
-} */
+@keyframes floatParticles {
+  0% {
+    background-position: 0px 0px;
+  }
+  100% {
+    background-position: 100px 100px;
+  }
+}
+
+
+@keyframes pulse {
+  0%, 100% {
+    transform: scale(1);
+  }
+  50% {
+    transform: scale(1.02);
+  }
+}
+
